requirements.txt
asgiref==3.11.1
attrs==26.1.0
coverage==7.13.5
Django==5.2.12
django-admin-rangefilter==0.13.5
django-debug-toolbar==6.3.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.9.0
mypy==1.20.2
mypy_extensions==1.1.0
packaging==26.1
pathspec==1.1.0
pillow==12.2.0
pluggy==1.6.0
psycopg2==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-cov==7.1.0
pytest-django==4.12.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0